PK 데이터타입 변경

+++ src/main/java/com/example/goENC/controllers/SurveyController.java
@@ -21,7 +21,7 @@ public class SurveyController {
     private final SurveyService surveyService;
 
     @PostMapping
-    public Integer createSurvey(@RequestBody RequestCreateSurveyDto requestDto) {
+    public Long createSurvey(@RequestBody RequestCreateSurveyDto requestDto) {
         return surveyService.createSurvey(requestDto);
     }
 
@@ -38,28 +38,28 @@ public class SurveyController {
 
     // 설문 ID에 따라 불러오기
     @GetMapping(value = "/{surveyId}")
-    public SurveyListResponseDto findBySurveyId(@PathVariable Integer surveyId) {
+    public SurveyListResponseDto findBySurveyId(@PathVariable Long surveyId) {
         return surveyService.findBySurveyId(surveyId);
     }
 
     // 설문 정보 업데이트하기
     @PutMapping(value = "/{surveyId}/meta")
-    public SurveyListResponseDto updateSurvey(@PathVariable Integer surveyId, @RequestBody SurveyUpdateDto surveyUpdateDto) {
+    public SurveyListResponseDto updateSurvey(@PathVariable Long surveyId, @RequestBody SurveyUpdateDto surveyUpdateDto) {
         return surveyService.updateSurvey(surveyId, surveyUpdateDto);
     }
 
     @PutMapping(value = "/{surveyId}/duplicate")
-    public Integer copyBySurveyId(@PathVariable Integer surveyId) {
+    public Long copyBySurveyId(@PathVariable Long surveyId) {
         return surveyService.copyBySurveyId(surveyId);
     }
 
     @PutMapping(value = "/{surveyId}/template")
-    public Integer reviseSurvey(@PathVariable Integer surveyId, @RequestBody RequestReviseSurveyDto requestDto) {
+    public Long reviseSurvey(@PathVariable Long surveyId, @RequestBody RequestReviseSurveyDto requestDto) {
         return surveyService.reviseSurvey(surveyId, requestDto);
     }
 
     @DeleteMapping(value = "/{surveyId}")
-    public Integer deleteFindBySurveyId(@PathVariable Integer surveyId) {
+    public Long deleteFindBySurveyId(@PathVariable Long surveyId) {
         return surveyService.deleteFindBySurveyId(surveyId);
     }
 }
